Move the signal setup down to ExecCommand()
aws-vault captures SIGINT in order to forward it to the spawned process.

Since the signal capturing is set up way before the chiled process is
spawn, in particular before getting the credentials, the signals are
captured at the mfa device prompt making it impossible to exit the CLI
using Ctrl-C at that point.

This moves the signal set up closer to when the child process is spawned,
making it possible to exit using Ctrl-C at the mfa prompt.