fix: format session key for filename compatibility
The session key has to be compatible with all Keyring backend stores,
which includes a File store that uses the key as a filename.

Colon is a disallowed filename character on Windows, so the separator
was changed to a comma.

KeyringSession fields have been updated to reflect the structured data
encoded in the session key, and the parser now only returns a populated
KeyringSession if the key was in the expected format.