Add support for AWS_MFA_SERIAL env var for rotate
Rotate didn't respect the AWS_MFA_SERIAL environment variable that is
used by the AWS SDK. when mfa_serial is defined in the profile directly,
or even 1 level of source_profile contains mfa_serial it made use of
that.
If however the mfa serial was provided as an environment variable with
AWS_MFA_SERIAL the rotate command failed.

This change allows rotate to pick up AWS_MFA_SERIAL env var and use a session
when the env var is defined.