refactor prompts to mfa-token-providers
Prompts were simple prompts, mfa token providers have the `mfa_serial`
that the token is being requested for. It may simply be used to display
to the user, or could be passed to the store in order to lookup the
correct creds. This enables the interfacing of apps/stores that manage
multiple mfa creds/secrets, examples of such stores may be bitwarden,
ykman.

The --prompt flag becomes --mfa-token-provider and AWS_VAULT_PROMPT is
changed to AWS_VAULT_MFA_TOKEN_PROVIDER.