Allow AWS region retrieval from metadata server
AWS region is returned via dynamic data, on instance identity document
endpoint of the metadata server. This region data is used by AWS SDK in
determining which region to consume the service on.

Allowing region retrieval on aws-vault server mode allows the server to
behave slightly more closely to the actual metadata server, and aids in
local application development, by allowing the code to be more agnostic
of its operating environment.