Fix for 99designs/aws-vault#408
Issue in kingpin cmdline parser around parsing partial shell completions.
Seems to have been mostly-partially fixed with alecthomas/kingpin#212, but never
pushed to gopkg.in.

This change moves to github.com, pinning to `master` branch/commit
at time of writing/test.

Also, fixes issue in `zsh` contributed completion; words want to start
at the second param and not be quoted.

There's still an issue if one has a profile name that's a substring of
a longer name (E.G. `one` and `one-ro`); will consider the matched
substring a full and unique match.  Completing the example, if two
profiles are named `one-rw` and `one-ro`, it will continue to match
both.