Set flags to cross-compile darwin/arm64 on darwin/amd64 and vice versa
This is an attempt to fix a lack of keychain support in darwin/arm64
binaries that have been cross-compiled on other platforms as described
in https://github.com/99designs/aws-vault/issues/758 and hinted at in
the linked
https://github.com/99designs/keyring/commit/756c48deecbd97e8fe3581f20efbaa136689cf80

Given the keychain support from keyring[1] is provided by cgo, and CGO
is disabled by default in cross-compilation, we need to enable that,
and deal with dev tooling/libraries.

I dug this solution from the Go issues, specifically
https://github.com/golang/go/issues/44112

Be warned, I am not familiar with the ins and outs of Go compilation,
especially when it comes to cross-compilation of CGO code, but at
least in this case, this change allows for a functional cross-compiled
binary.

I fully expect that attempting to cross-compile darwin/arm64 on
anything other than darwin/amd64 (or the opposite way around) is going
to end badly.

[1] https://github.com/99designs/keyring