Add OpenSUSE update instructions
The aws-vault software is packaged and built for OpenSUSE/SUSE, but the version update process is manual at the moment.  It would be fairly easy to set it to auto-build new versions, but that would require someone to enable a webhook on the repo which notifies the OpenBuildService when a new version is released.  I'd be happy to help set that up on the SUSE side if the repo owner is willing to add the hook and create an OBS account.  Otherwise, I suppose I could set up my own fork and add a scheduled action to sync mine from upstream.