Web identity profiles with role ARNs are valid
Profiles that specify web identity process or files require the role ARN to be
set as well. Since the 7.0.0 release, this has been failing validation because
it is now considered to have two sources of credentials. This fixes the
validation to correctly consider this as a profile with one source credential.

Fixes #1177